ability to deploy javascript code on chain to a contract
currently, it is required to include the code to be executed inline when requesting functions execution. this is great,
as it means that the code to be executed has to be on-chain to be trusted. however, as contracts and their corresponding
offchain code gets more complicated, it becomes necesary to store the code in a separate contract in order to stay
under the EVM size limit. Fortunately, it is actually possible to push a non-functional contract embedde with the blob
of the code directly on-chain. Its a reaonsably efficient way to do so as a execution node (for ethereum or otherwise) is
able to store contract source data contiguously in storage, allowing for efficient load by `extcode` or whatever.

This PR proposed addition of a function which will allow for the specified javascript functions file to be uploaded on-chain
as a contract. the bytecode of the contract will effectively just be the contents of the javascript file, and these can be
loaded by a Functions Consumer for inline execution by a function similar to the following:

```

    /**
     * loads the bytes deployed to the specified address
     * used to get the inline execution code for chainlink
     */
    function _codeAt(address _addr) public view returns (bytes o_code) {
        assembly {
            // retrieve the size of the code, this needs assembly
            let size := extcodesize(_addr)
            // allocate output byte array - this could also be done without assembly
            // by using o_code = new bytes(size)
            o_code := mload(0x40)
            // new "memory end" including padding
            mstore(0x40, add(o_code, and(add(add(size, 0x20), 0x1f), not(0x1f))))
            // store length in memory
            mstore(o_code, size)
            // actually retrieve the code, this needs assembly
            extcodecopy(_addr, add(o_code, 0x20), 0, size)
        }
    }
```

(alternatively, functions could instead encode the address of the code into the CBOR and save the
consumer 10k gas or so per call, but one step at a time :laugh:)

this same pattern could be applied to secrets as well.